# Replace stdlib CSV reader with simpler detector