--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,4 @@ module github.com/gabriel-vasile/mimetype
 
 go 1.20
 
-require golang.org/x/net v0.25.0
+require golang.org/x/net v0.33.0