There is evidence that using the stdlib csv reader can be resource intensive from a memory perspective:
- https://github.com/gabriel-vasile/mimetype/issues/354
- https://github.com/golang/go/issues/8059
- https://github.com/golang/go/issues/16786
- https://github.com/golang/go/issues/20169

We're seeing evidence of this in [stereoscope](https://github.com/anchore/stereoscope):

<img width="1060" alt="Screenshot 2024-07-10 at 12 28 47 PM" src="https://github.com/anchore/mimetype/assets/590471/3d866dda-64f3-40ab-9e4b-1801293ecaf7">

Since we are not in need of the full CSV reader functionality, this PR drops usage of the CSV reader and adds a CSV detector in its place. This yields a drastic performance improvement memory-wise (not inuse memory, total allocated memory):

<img width="1060" alt="Screenshot 2024-07-10 at 12 29 38 PM" src="https://github.com/anchore/mimetype/assets/590471/ac492ee4-1aff-432a-bae3-a28e163aa249">