@wagoodman 2024-07-15
I'll push the refactors shortly, but I wrote a benchmark test locally that I wont push since the `sv` code will be gone.

<details>
<summary>Benchmark test code</summary>

```golang
func BenchmarkDetectVsSv(b *testing.B) {
	fh, err := os.Open("random_data.csv")
	if err != nil {
		b.Fatalf("failed to open file: %+v", err)
	}

	contents, err := io.ReadAll(fh)
	if err != nil {
		b.Fatalf("failed to read file: %+v", err)
	}

	for _, limit := range []uint32{0, 100, 1000} {
		b.Run(fmt.Sprintf("sv(limit=%d)", limit), func(b *testing.B) {
			sv(contents, ',', limit)
		})

		b.Run(fmt.Sprintf("Detect(limit=%d)", limit), func(b *testing.B) {
			Detect(contents, ',', limit)
		})
	}

}

```

</details>

```
BenchmarkDetectVsSv
BenchmarkDetectVsSv/sv(limit=0)
BenchmarkDetectVsSv/sv(limit=0)-12         	1000000000	         0.001642 ns/op
BenchmarkDetectVsSv/Detect(limit=0)
BenchmarkDetectVsSv/Detect(limit=0)-12     	1000000000	         0.0000238 ns/op
BenchmarkDetectVsSv/sv(limit=100)
BenchmarkDetectVsSv/sv(limit=100)-12       	1000000000	         0.001492 ns/op
BenchmarkDetectVsSv/Detect(limit=100)
BenchmarkDetectVsSv/Detect(limit=100)-12   	1000000000	         0.0000017 ns/op
BenchmarkDetectVsSv/sv(limit=1000)
BenchmarkDetectVsSv/sv(limit=1000)-12      	1000000000	         0.001519 ns/op
BenchmarkDetectVsSv/Detect(limit=1000)
BenchmarkDetectVsSv/Detect(limit=1000)-12  	1000000000	         0.0000065 ns/op
PASS
```

So `Detect()` is winning out in terms of CPU time too (🎉 🌮 ).

The file under test was:
- 421K in file size
- 40 columns per row
- 1500 rows
- mixed number and strings field types
- strings have a mix of single and double quote usage