@wagoodman 2024-07-15
After running through the fuzz test, I've found a lot of edge cases. It essentially comes down to the `LazyQuotes` setting on the original CSV reader. I'll see what I can do, but in the meantime I've converted this to a draft to signal it might be a little while before it's ready for review...