@wagoodman 2024-07-17
@gabriel-vasile I've done some initial work to get the prototype working with fuzz tests, but honestly it's not worth the gains (less memory performance improvement, worse performance CPU-wise, crazy complex). The current state of the prototype is here https://github.com/anchore/mimetype/pull/2 ... quote larger than this current PR.

I might shelve this for a while, but will noodle on it in the background. One thought I had: what if using the "simple" (more performant detector) that did not support LazyQuotes as a configuration option? I also haven't characterized just switching this flag off when creating the stdlib reader.