@gabriel-vasile 2024-08-12
@wagoodman I took inspiration from [json package](https://github.com/golang/go/blob/2f3fee058f3bb1c98b8ecc5aeb2329aaadf4dc0b/src/encoding/json/scanner.go#L87-L107) and I added a pool of `bufio.Readers` for CSV and TSV detection.

The code is here: 5f825db8cf4265eada32fe625781cdb75193e6f7

The benchmark shows -77% allocated bytes:
```
➜  magic git:(csvpprof) ✗ benchstat before after
goos: linux
goarch: amd64
pkg: github.com/gabriel-vasile/mimetype/internal/magic
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
      │   before    │               after                │
      │   sec/op    │   sec/op     vs base               │
Csv-8   3.052µ ± 3%   1.695µ ± 3%  -44.46% (p=0.001 n=7)

      │    before    │                after                │
      │     B/op     │     B/op      vs base               │
Csv-8   5.258Ki ± 0%   1.164Ki ± 0%  -77.86% (p=0.001 n=7)

      │   before   │               after               │
      │ allocs/op  │ allocs/op   vs base               │
Csv-8   13.00 ± 0%   11.00 ± 0%  -15.38% (p=0.001 n=7)
```

It's not ideal, there are still allocations done, but it cuts the bulk of allocated bytes. Let me know if you see improvements in stereoscope, but this commit is probably going to master.