@gabriel-vasile 2025-08-25
https://github.com/gabriel-vasile/mimetype/pull/672 made some more improvements to CSV detection. Allocations should be reduced now. I'll close this.